feat: user can specify tags
